Test with unambiguous doubles.
The next release of readr (1.1.0) modifies how numeric vectors
containing whole numbers are written by write_* functions. Previous
these numbers were written with a trailing `.0`, which ensures they were
always read as numerics as well.

This behavior was a change from R's typical output (which prints without
a trailing `.0`), so readr has been changed to do the same.

This change causes a test failure, as when the numbers are read back in
they are read as integers rather than numerics. Changing the test data
to use unambiguous double values fixes the issue.

This PR also adds a `...` argument to MergeTablesOnDisk, which allows
users to pass additional arguments to `readr::read_delim` if necessary
(such as supplying explicit column types).